refactor: Migrate Prime purchase web embed to use RevenueCat Purchases SDK directly
